Move swap to swap_network.
