+++ qualtran/bloqs/swap_network_test.py
@@ -21,19 +21,23 @@ import pytest
 import sympy
 
 import qualtran.cirq_interop.testing as cq_testing
-from qualtran import Bloq, BloqBuilder
+from qualtran import Bloq, BloqBuilder, SelectionRegister
 from qualtran.bloqs.basic_gates import CSwap, TGate
 from qualtran.bloqs.basic_gates.z_basis import IntState
 from qualtran.bloqs.swap_network import (
     _approx_cswap_large,
     _approx_cswap_small,
     _approx_cswap_symb,
+    _multiplexed_cswap,
     _swz_small,
     CSwapApprox,
+    MultiplexedCSwap,
     SwapWithZero,
 )
 from qualtran.bloqs.util_bloqs import ArbitraryClifford
+from qualtran.cirq_interop.bit_tools import iter_bits
 from qualtran.cirq_interop.t_complexity_protocol import TComplexity
+from qualtran.cirq_interop.testing import assert_circuit_inp_out_cirqsim, GateHelper
 from qualtran.simulation.quimb_sim import flatten_for_tensor_contraction
 from qualtran.testing import assert_valid_bloq_decomposition, execute_notebook
 
@@ -203,6 +207,63 @@ def test_swap_with_zero_bloq_counts(selection_bitsize, target_bitsize, n_target_
 def test_swap_with_zero_t_complexity(selection_bitsize, target_bitsize, n_target_registers, want):
     gate = SwapWithZero(selection_bitsize, target_bitsize, n_target_registers)
     assert want == gate.t_complexity()
+
+
+@pytest.mark.parametrize(
+    "selection_bitsize,iteration_length,target_bitsize", [[2, 3, 2], [3, 2, 3]]
+)
+def test_cswap_lth_reg(selection_bitsize, iteration_length, target_bitsize):
+    greedy_mm = cirq.GreedyQubitManager(prefix="_a", maximize_reuse=True)
+    gate = MultiplexedCSwap(
+        SelectionRegister('selection', selection_bitsize, iteration_length),
+        target_bitsize=target_bitsize,
+    )
+    g = GateHelper(gate, context=cirq.DecompositionContext(greedy_mm))
+    for n in range(iteration_length):
+        # Initial qubit values
+        qubit_vals = {q: 0 for q in g.all_qubits}
+        # Set selection according to `n`
+        qubit_vals.update(zip(g.quregs['selection'], iter_bits(n, selection_bitsize)))
+        final_state = [qubit_vals[x] for x in g.all_qubits]
+
+        # swap the nth register (x{n}) with the ancilla (y)
+        # put some non-zero numbers in the registers for comparison.
+        qubit_vals.update(zip(g.quregs['targets'][n], iter_bits(n + 1, target_bitsize)))
+        initial_state = [qubit_vals[x] for x in g.all_qubits]
+        qubit_vals.update(zip(g.quregs['targets'][n], [0] * len(g.quregs['targets'][n])))
+        qubit_vals.update(zip(g.quregs['output'], iter_bits(n + 1, target_bitsize)))
+        final_state = [qubit_vals[x] for x in g.all_qubits]
+        assert_circuit_inp_out_cirqsim(
+            g.decomposed_circuit, g.all_qubits, initial_state, final_state
+        )
+
+
+@pytest.mark.parametrize(
+    "selection_bitsize,iteration_length,target_bitsize", [[2, 3, 2], [3, 2, 3]]
+)
+def test_bloq_has_consistent_decomposition(selection_bitsize, iteration_length, target_bitsize):
+    bloq = MultiplexedCSwap(
+        SelectionRegister('selection', selection_bitsize, iteration_length),
+        target_bitsize=target_bitsize,
+    )
+    assert_valid_bloq_decomposition(bloq)
+
+
+@pytest.mark.parametrize(
+    "selection_bitsize,iteration_length,target_bitsize", [[3, 8, 2], [4, 9, 3]]
+)
+def test_t_counts(selection_bitsize, iteration_length, target_bitsize):
+    bloq = MultiplexedCSwap(
+        SelectionRegister('selection', selection_bitsize, iteration_length),
+        target_bitsize=target_bitsize,
+    )
+    expected = 4 * (iteration_length - 2) + 7 * (iteration_length * target_bitsize)
+    assert bloq.t_complexity().t == expected
+    assert bloq.call_graph()[1][TGate()] == expected
+
+
+def test_multiplexed_cswap(bloq_autotester):
+    bloq_autotester(_multiplexed_cswap)
 
 
 def test_approx_cswap_small(bloq_autotester):
